Interview_Prep/Zeal/Code_Challenge/requirements.txt
confluent-kafka==2.6.0
psycopg2-binary==2.9.9

